# libs/contracts/transport/http/swagger/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST service of online subscriptions at present moment ()",
        "title": "API-Gateway",
        "contact": {
            "name": "Yan",
            "email": "89269167763@yandex.ru.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/subscriptions": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Create subscription",
                "parameters": [
                    {
                        "description": "Payload to create subscription",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/customerHttpContract.CustomerSubscriptionCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/customerHttpContract.CustomerSubscriptionCreateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpContractCommons.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpContractCommons.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/subscriptions/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Get subscription by id",
                "parameters": [
                    {
                        "type": "string",
                        "example": "60601fee-2bf1-4721-ae6f-7636e79a0cba",
                        "description": "subscription ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/customerHttpContract.CustomerSubscriptionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpContractCommons.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpContractCommons.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpContractCommons.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Delete subscription by ID",
                "parameters": [
                    {
                        "type": "string",
                        "example": "4b70f8d6-c702-4e6e-9c65-2ae0e3bb0e5c",
                        "description": "subscription ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "no content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpContractCommons.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpContractCommons.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Update subscription by ID",
                "parameters": [
                    {
                        "type": "string",
                        "example": "4b70f8d6-c702-4e6e-9c65-2ae0e3bb0e5c",
                        "description": "subscription ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/customerHttpContract.CustomerSubscriptionUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "no content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpContractCommons.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpContractCommons.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "customerHttpContract.CustomerSubscriptionCreateRequest": {
            "type": "object",
            "required": [
                "price",
                "service_name",
                "start_date",
                "user_id"
            ],
            "properties": {
                "end_date": {
                    "description": "@example \"07-2026\"",
                    "type": "string"
                },
                "price": {
                    "description": "@example 400",
                    "type": "integer"
                },
                "service_name": {
                    "description": "@example \"Yandex Plus\"",
                    "type": "string"
                },
                "start_date": {
                    "description": "@example \"07-2025\"",
                    "type": "string"
                },
                "user_id": {
                    "description": "@example \"60601fee-2bf1-4721-ae6f-7636e79a0cba\"",
                    "type": "string"
                }
            }
        },
        "customerHttpContract.CustomerSubscriptionCreateResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "customerHttpContract.CustomerSubscriptionResponse": {
            "type": "object",
            "properties": {
                "end_date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "service_name": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "customerHttpContract.CustomerSubscriptionUpdateRequest": {
            "type": "object",
            "properties": {
                "end_date": {
                    "description": "@example \"07-2026\"",
                    "type": "string"
                },
                "price": {
                    "description": "@example 400",
                    "type": "integer"
                },
                "service_name": {
                    "description": "@example \"Yandex Plus\"",
                    "type": "string"
                },
                "start_date": {
                    "description": "@example \"07-2025\"",
                    "type": "string"
                }
            }
        },
        "httpContractCommons.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}